Automatically backed up by Learn

+++ current_age_for_birth_year.rb
@@ -1,4 +1,4 @@
-age_of_person = 2003 - current_age_for_birth_year.to_f
+age_of_person = 2003 - current_age_for_birth_year.to_i
 def current_age_for_birth_year
   puts "You are #{age_of_person} years old"
 end
